requirements.txt
discord==2.2.0
numpy==1.26.4
pydantic==2.3
markdownify==0.11.6
matplotlib==3.6.3
beautifulsoup4
python-dotenv
pymongo
urllib3
tldextract==5.1.2
factordb-python